requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dj-database-url==3.0.1
Django==5.1.5
django-paypal==2.1
gunicorn==23.0.0
idna==3.10
joblib==1.4.2
mssql-django==1.6
nltk==3.9.1
packaging==25.0
paypalrestsdk==1.13.3
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
pygame==2.6.1
pyodbc==5.2.0
pyOpenSSL==25.0.0
pytz==2025.2
razorpay==1.4.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
whitenoise==6.10.0
